Get double for number

+++ src/android-native/src/main/java/io/liteglue/SQLitePluginConverter.java
@@ -35,7 +35,7 @@ public abstract class SQLitePluginConverter {
                     jsonObject.put(key, readableMap.getBoolean(key));
                     break;
                 case Number:
-                    jsonObject.put(key, readableMap.getInt(key));
+                    jsonObject.put(key, readableMap.getDouble(key));
                     break;
                 case String:
                     jsonObject.put(key, readableMap.getString(key));
@@ -64,7 +64,7 @@ public abstract class SQLitePluginConverter {
                     jsonArray.put(readableArray.getBoolean(i));
                     break;
                 case Number:
-                    jsonArray.put(readableArray.getInt(i));
+                    jsonArray.put(readableArray.getDouble(i));
                     break;
                 case String:
                     jsonArray.put(readableArray.getString(i));
